Only save draft when the form values change.
There is a race condition between when async draft saving starts and when useFormSync renders, such that this effect can run again before loading becomes true.

This causes the draft saving logic to run again, but by that point the document has already been updated. This causes loadingDraft to be dispatched in useEditTestimony, but not loadDraft, because the snapshot listener only runs when the document actually changes.

The fix here is not to try to save the draft unless the form values actually change.