test(e2e): fix timeout error when clicking create testimony
Fixed an issue where the tests would time out when clicking the 'Create Testimony' button in the submit testimony test. Now that I am using getByRole to find the button instead of waitForSelector, all browsers except firefox pass this part of the test. Firefox fails waiting for profile icon in the beforeEach block.